Add missing chapters.

+++ vector/vector.c
@@ -6,7 +6,7 @@ static void grow(struct hc_vector *v) {
   hc_vector_grow(v, v->capacity ? v->capacity*2 : 2);
 }
 
-struct hc_vector *hc_vector_init(struct hc_vector *v, int item_size) {
+struct hc_vector *hc_vector_init(struct hc_vector *v, size_t item_size) {
   v->item_size = item_size;
   v->capacity = v->length = 0;
   v->items = v->start = v->end = NULL;
@@ -17,7 +17,7 @@ void hc_vector_deinit(struct hc_vector *v) {
   if (v->items) { free(v->items); }
 }
 
-void hc_vector_grow(struct hc_vector *v, int capacity) {
+void hc_vector_grow(struct hc_vector *v, size_t capacity) {
   v->capacity = capacity;
 
   v->items = realloc(v->items,
@@ -32,7 +32,7 @@ void hc_vector_clear(struct hc_vector *v) {
   v->end = v->start;
 }
 
-void *hc_vector_get(struct hc_vector *v, int i) {
+void *hc_vector_get(struct hc_vector *v, size_t i) {
   return v->items ? v->start + v->item_size*i : NULL;
 }
 
@@ -55,8 +55,8 @@ void *hc_vector_pop(struct hc_vector *v) {
   return v->end;
 }
 
-void *hc_vector_insert(struct hc_vector *v, int i, int n) {
-  const int m = v->length+n;
+void *hc_vector_insert(struct hc_vector *v, size_t i, size_t n) {
+  const size_t m = v->length+n;
   if (m > v->capacity) { hc_vector_grow(v, m); } 
   uint8_t *const p = hc_vector_get(v, i);
 
@@ -69,8 +69,8 @@ void *hc_vector_insert(struct hc_vector *v, int i, int n) {
   return p;
 }
 
-bool hc_vector_delete(struct hc_vector *v, int i, int n) {
-  const int m = i+n;
+bool hc_vector_delete(struct hc_vector *v, size_t i, size_t n) {
+  const size_t m = i+n;
   if (v->length < m) { return false; }
 
   if (m < v->length) {
